针对登录添加了Gateway网关和Jwt无状态加密。此外为确保服务的高可用性，采用Sentinel用于服务的限流、降级、熔断。

+++ notes/src/main/java/com/example/notes/dto/NoteDTO.java
@@ -40,7 +40,7 @@ public class NoteDTO {
     /** 点赞数 */
     private Integer likeNum;
     /** 收藏数 */
-    private Double collectionsNum;
+    private Integer collectionsNum;
     /** 更新时间 */
     private Timestamp updateTime;
     /** 创建时间 */
